fixed the P.R. suggestion
